requirements.txt
Flask==2.2.2
google-cloud-storage==2.5.0
google-cloud-firestore==2.5.0
tensorflow==2.11.0
numpy==1.23.5
pandas==1.5.3
